refactor: FramesOfPlayer 클래스를 Player 클래스로 기능 이임 후 삭제
